Reorganise slightly
