+++ openfisca_us/variables/demographic/tax_unit.py
@@ -2,18 +2,34 @@ from openfisca_core.model_api import *
 from openfisca_us.entities import *
 
 
-class person_id(Variable):
-    value_type = int
-    entity = Person
-    label = u"Unique reference for this person"
+class tax_unit_id(Variable):
+    value_type = float
+    entity = TaxUnit
+    label = u"Unique reference for this tax unit"
     definition_period = ETERNITY
 
 
-class person_weight(Variable):
+class tax_unit_weight(Variable):
     value_type = float
-    entity = Person
-    label = u"Person weight"
+    entity = TaxUnit
+    label = u"Tax unit weight"
     definition_period = YEAR
+
+
+class person_tax_unit_id(Variable):
+    value_type = int
+    entity = Person
+    label = u"Unique reference for the tax unit of this person"
+    definition_period = ETERNITY
+
+
+class MARSType(Enum):
+    SINGLE = "Single"
+    JOINT = "Joint"
+    SEPARATE = "Separate"
+    HOUSEHOLD_HEAD = ("Head of household",)
+    WIDOW = "Widow(er)"
+
 
 
 class is_tax_unit_head(Variable):
@@ -49,10 +65,3 @@ class is_tax_unit_dependent(Variable):
             person("is_tax_unit_head", period)
             + person("is_tax_unit_spouse", period)
         )
-
-
-class age(Variable):
-    value_type = int
-    entity = Person
-    label = u"Age"
-    definition_period = YEAR
